# dev/ci: only asdf install go in pipeline gen