--- a/.buildkite/hooks/pre-command
+++ b/.buildkite/hooks/pre-command
@@ -23,8 +23,12 @@ mkdir -p ./annotations/
 
 # asdf setup
 # ----------
-if [[ ! "$BUILDKITE_STEP_KEY" == "pipeline-upload" ]]; then
-  ./dev/ci/asdf-install.sh
+if [[ "$BUILDKITE_STEP_KEY" == "pipeline-gen" ]]; then
+  echo "running go install only"
+  asdf install golang
+elif [[ "$BUILDKITE_STEP_KEY" == "pipeline-upload" ]]; then
+  echo "skipping asdf install for pipeline upload step"
 else
-  echo "skipping asdf install for pipeline upload setp"
+  echo "running normal install"
+  ./dev/ci/asdf-install.sh
 fi