--- a/.buildkite/hooks/pre-command
+++ b/.buildkite/hooks/pre-command
@@ -23,11 +23,13 @@ mkdir -p ./annotations/
 
 # asdf setup
 # ----------
+echo "~~~ Preparing asdf dependencies"
+
 if [[ "$BUILDKITE_STEP_KEY" == "pipeline-gen" ]]; then
-  echo "running go install only"
+  echo "pipeline-gen step: running go install only"
   asdf install golang
 elif [[ "$BUILDKITE_STEP_KEY" == "pipeline-upload" ]]; then
-  echo "skipping asdf install for pipeline upload step"
+  echo "pipeline-upload step: skipping asdf install"
 else
   echo "running normal install"
   ./dev/ci/asdf-install.sh