Cuts the pipeline generation step from ~60s to ~25s - doing a fresh `asdf install golang` seems significantly faster than using the asdf tools cache.

Opened a related idea: https://github.com/sourcegraph/sourcegraph/issues/33854

Together with https://github.com/sourcegraph/sourcegraph/pull/33821 , this significantly reduces the time it takes for builds to get to the "actually running" state (~120s -> ~50s)

## Test plan

<!--
  As part of SOC2/GN-104 and SOC2/GN-105 requirements, all pull requests are REQUIRED to
  provide a "test plan". A test plan is a loose explanation of what you have done or
  implemented to test this, or why this change does not need testing, as outlined in our
  Testing principles and guidelines:
  https://docs.sourcegraph.com/dev/background-information/testing_principles
  Write your test plan here after the "## Test plan" header.
-->

https://buildkite.com/sourcegraph/sourcegraph/builds/142131